refactor: moved biconomySponsoredPaymasterContext to paymaster client file
